De-duplicate arrowhead code
